+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### On going
+* Fix scroll actions
+
 ### 0.3.2: May 11th, 2017
 * Fix password protected form action